--- a/benchmarks/benchmark.py
+++ b/benchmarks/benchmark.py
@@ -4,11 +4,10 @@ import pysbd
 import spacy
 import stanza
 
-import syntok
 from syntok.tokenizer import Tokenizer
 import syntok.segmenter as syntok_segmenter
 
-from english_golden_rules import GOLDEN_EN_RULES
+from .english_golden_rules import GOLDEN_EN_RULES
 
 pysbd_segmenter = pysbd.Segmenter(language="en", clean=False, char_span=False)
 
@@ -62,6 +61,15 @@ def benchmark(golden_rules, tokenize_func):
 
     return percent_score
 
+
+def speed_benchmark(filepath, tokenize_func):
+    sentences = []
+    for line in open(filepath):
+        segments = tokenize_func(line)
+        sentences.append(segments)
+    return sentences
+
+
 if __name__ == "__main__":
     import time
     libraries = (
@@ -74,11 +82,14 @@ if __name__ == "__main__":
         syntok_tokenize)
     for tokenize_func in libraries:
         t = time.time()
-        for i in range(100):
-            percent_score = benchmark(GOLDEN_EN_RULES, tokenize_func)
+        # wget http://www.gutenberg.org/files/1661/1661-0.txt -P benchmarks/
+        sentences = speed_benchmark('benchmarks/1661-0.txt', tokenize_func)
+        # for i in range(100):
+        #     percent_score = benchmark(GOLDEN_EN_RULES, tokenize_func)
 
         time_taken = time.time() - t
         print()
         print(tokenize_func.__name__)
-        print('GRS score: {:0.2f}%'.format(percent_score))
-        print('Speed(Avg over 100 runs): {:>10.2f} ms'.format(time_taken*1000/100))
+        print('Speed : {:>20.2f} ms'.format(time_taken * 1000))
+        # print('GRS score: {:0.2f}%'.format(percent_score))
+        # print('Speed(Avg over 100 runs): {:>10.2f} ms'.format(time_taken*1000/100))