--- a/pysbd/utils.py
+++ b/pysbd/utils.py
@@ -6,7 +6,7 @@ import pysbd
 class Rule(object):
 
     def __init__(self, pattern, replacement):
-        self.pattern = re.compile(pattern)
+        self.pattern = re.compile(pattern)  # pragma: no cover
         self.replacement = replacement
 
     def __repr__(self):  # pragma: no cover