As mentioned in #41 , `abbreviation_replacer.py` takes too long and needs to be refactored and needs a performance improvement.

# Speed Benchmark on bigger text file


Tool | Speed
-- | --
blingfire_tokenize | 55.63 ms
nltk_tokenize | 198.17 ms
pysbd_tokenize | 12846.23 ms
spacy_tokenize | 741.54 ms
spacy_dep_tokenize | 17642.21 ms
stanza_tokenize | 35623.08 ms
syntok_tokenize | 1455.21 ms

Text file used: http://www.gutenberg.org/files/1661/1661-0.txt
> wget http://www.gutenberg.org/files/1661/1661-0.txt -P benchmarks/